requirements.txt
yt-dlp==2024.3.10
python-dotenv==1.0.1
pymongo==4.7.2
minio==7.2.5
azure-storage-blob==12.19.1
requests==2.31.0 